Add JSON storage support.
The version was upgraded to 0.3.0: 0.2.0 was a locale update,
this one is JSON update.

This update introduced a StorageMethod class that is responsible
for all the current and future storage implementations —
serialization, JSON and further SQL storage.

Implementations are provided through the StorageFactory class.

The JSON storage implementation lies in the storage.json package
and it's rather simple implementation using the provided Guava
json library.

The new configuration property was added to the config.yml file
that is now responsible for the storage method specification.